refactor(sim_codegen): extract gen_fifo into submodule
